+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Renamed `delineate_intersecting_shards` to `dedupe_policy_shards` to better reflect how people will use it.
 - Added `explain_policy_shards` to eventually replace `explain` attribute on `PolicyShard` class entirely.
 - Added `__contains__` to `EffectiveARP` classes
-
+- Added `reverse` method to `Condition` to reverse the operator/value to reverse the effect of the condition.
 # 0.5.0
 
 - Renamed `dedupe_policy_shards` to `dedupe_policy_shard_subsets` to differentiate it from `delineate_intersecting_shards`.